V1.7.07 - Updates
- Eliminated all warnings under PlatformIO compilation for all boards

+++ Software/Arduino code/OpenAstroTracker/WifiControl.hpp
@@ -2,8 +2,8 @@
 #include "Globals.hpp"
 
 #ifdef WIFI_ENABLED
-#include "Mount.hpp";
-#include "LcdMenu.hpp";
+#include "Mount.hpp"
+#include "LcdMenu.hpp"
 #include "MeadeCommandProcessor.hpp"
 #include "WiFiServer.h"
 #include "WiFiUDP.h"
@@ -39,8 +39,8 @@ private:
     WiFiUDP* _udp;
     WiFiClient client;
 
-    long _infraStart = 0;
-    int _infraWait = 30000; // 30 second timeout for 
+    unsigned long _infraStart = 0;
+    unsigned long _infraWait = 30000; // 30 second timeout for 
 };
 
 extern WifiControl wifiControl;
